+++ examples/__tests__/integration/koa.test.ts
@@ -1,0 +1,4 @@
+import { deployExample } from '../test-utils';
+it('[examples] should deploy koa', async () => {
+  await deployExample('koa');
+});